kaliyai-rag/requirements.txt
# Kaliyai RAG ingestion & retrieval helpers (Python 3.11+)
httpx>=0.27,<1
pydantic>=2.6,<3
python-dotenv>=1.0,<2
PyYAML>=6.0,<7
trafilatura>=2.0,<3
# PDF → text (Tesseract OCR); needs system `tesseract` + Poppler (`pdftoppm`)
pdf2image>=1.17,<2
pytesseract>=0.3.10,<1
Pillow>=10,<11
pypdf>=5,<6

# Optional: pytest -- kaliyai-rag/tests
pytest>=8,<9
